Moved some files, changed some namespaces.
