+++ Domain/WindowTree/Helpers/BoundedTreeNode.cs
@@ -1,8 +1,8 @@
-﻿using Kaboom.Domain.WindowTree.General;
+﻿using Kaboom.Domain.WindowTree.ValueObjects;
 using System;
 using System.Collections.Generic;
 
-namespace Kaboom.Domain.WindowTree.ArrangementAggregate
+namespace Kaboom.Domain.WindowTree.Helpers
 {
     public abstract class BoundedTreeNode : IBoundedTreeNode
     {
