+++ Domain/WindowTree/Helpers/IBoundedTreeNode.cs
@@ -1,7 +1,7 @@
 ﻿using Kaboom.Abstraction;
-using Kaboom.Domain.WindowTree.General;
+using Kaboom.Domain.WindowTree.ValueObjects;
 
-namespace Kaboom.Domain.WindowTree.ArrangementAggregate
+namespace Kaboom.Domain.WindowTree.Helpers
 {
     public interface IBoundedTreeNode : ITreeNode<IBoundedTreeNode>, IEntity
     {
