+++ Domain/WindowTree/IArrangementRepository.cs
@@ -1,7 +1,7 @@
-﻿using Kaboom.Domain.WindowTree.General;
+﻿using Kaboom.Domain.WindowTree.ValueObjects;
 using System.Collections.Generic;
 
-namespace Kaboom.Domain.WindowTree.ArrangementAggregate
+namespace Kaboom.Domain.WindowTree
 {
     public interface IArrangementRepository
     {
