+++ Domain/WindowTree/ValueObjects/Direction.cs
@@ -1,4 +1,4 @@
-﻿namespace Kaboom.Domain.WindowTree.General
+﻿namespace Kaboom.Domain.WindowTree.ValueObjects
 {
     public class Direction
     {
