+++ Testing/Application/ActionTests.cs
@@ -1,6 +1,6 @@
 ﻿using Kaboom.Application;
 using Kaboom.Application.Actions.WorkspaceActions;
-using Kaboom.Domain.WindowTree.General;
+using Kaboom.Domain.WindowTree.ValueObjects;
 using Kaboom.Testing.Mocks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
