+++ Testing/Domain/ArrangementTests.cs
@@ -1,5 +1,7 @@
-﻿using Kaboom.Domain.WindowTree.ArrangementAggregate;
+﻿using Kaboom.Domain.WindowTree;
 using Kaboom.Domain.WindowTree.General;
+using Kaboom.Domain.WindowTree.Helpers;
+using Kaboom.Domain.WindowTree.ValueObjects;
 using Kaboom.Testing.Mocks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
