+++ Testing/Domain/WindowTests.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Kaboom.Domain.WindowTree.Exceptions;
-using Kaboom.Domain.WindowTree.ArrangementAggregate;
 using Kaboom.Testing.Mocks;
-using Kaboom.Domain.WindowTree.General;
+using Kaboom.Domain.WindowTree.ValueObjects;
+using Kaboom.Domain.WindowTree;
 
 namespace Kaboom.Testing.Domain
 {
