+++ Testing/Mocks/MockArrangement.cs
@@ -1,5 +1,7 @@
-﻿using Kaboom.Domain.WindowTree.ArrangementAggregate;
-using Kaboom.Domain.WindowTree.General;
+﻿using Kaboom.Domain;
+using Kaboom.Domain.WindowTree;
+using Kaboom.Domain.WindowTree.Helpers;
+using Kaboom.Domain.WindowTree.ValueObjects;
 using System.Collections.Generic;
 
 namespace Kaboom.Testing.Mocks
