+++ Testing/Mocks/MockArrangementRepository.cs
@@ -1,5 +1,6 @@
-﻿using Kaboom.Domain.WindowTree.ArrangementAggregate;
-using Kaboom.Domain.WindowTree.General;
+﻿using Kaboom.Domain;
+using Kaboom.Domain.WindowTree;
+using Kaboom.Domain.WindowTree.ValueObjects;
 using System.Collections.Generic;
 using System.Linq;
